Fixing connection config.
