+++ crates/utils/src/settings/mod.rs
@@ -5,10 +5,14 @@ use regex::Regex;
 use std::{env, fs, sync::LazyLock};
 use structs::{PictrsConfig, Settings};
 use url::Url;
+use urlencoding::encode;
 
 pub mod structs;
 
 static DEFAULT_CONFIG_FILE: &str = "config/config.hjson";
+
+/// Some connection options to speed up queries
+const CONNECTION_OPTIONS: [&str; 1] = ["geqo_threshold=12"];
 
 #[allow(clippy::expect_used)]
 pub static SETTINGS: LazyLock<Settings> = LazyLock::new(|| {
@@ -97,6 +101,30 @@ impl Settings {
       .pictrs
       .clone()
       .ok_or_else(|| anyhow!("images_disabled").into())
+  }
+
+  /// Sets a few additional config options necessary for starting lemmy
+  pub fn get_database_url_with_options(&self) -> LemmyResult<String> {
+    let mut url = Url::parse(&self.get_database_url())?;
+
+    // Set `lemmy.protocol_and_hostname` so triggers can use it
+    let lemmy_protocol_and_hostname_option =
+      "lemmy.protocol_and_hostname=".to_owned() + &self.get_protocol_and_hostname();
+    let mut options = CONNECTION_OPTIONS.to_vec();
+    options.push(&lemmy_protocol_and_hostname_option);
+
+    // Create the connection uri portion
+    let options_segments = options
+      .iter()
+      // The equal signs need to be encoded, since the url set_query doesn't do them,
+      // and postgres requires them to be %3D
+      // https://www.postgresql.org/docs/current/libpq-connect.html#LIBPQ-CONNSTRING
+      .map(|o| format!("-c {}", encode(&o)))
+      .collect::<Vec<String>>()
+      .join(" ");
+
+    url.set_query(Some(&format!("options={options_segments}")));
+    Ok(url.into())
   }
 }
 #[allow(clippy::expect_used)]
